+++ src/TrelloDotNet/Model/TrelloApiException.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Net;
 
 namespace TrelloDotNet.Model
 {
@@ -11,6 +12,11 @@ namespace TrelloDotNet.Model
         /// What URL was sent to Trello (good to debug with in Postman and similar tools)
         /// </summary>
         public string DataSentToTrello { get; }
+
+        /// <summary>
+        /// The underlying HTTP Status code
+        /// </summary>
+        public HttpStatusCode? ErrorCode { get; }
 
         /// <summary>
         /// Constructor
@@ -26,9 +32,11 @@ namespace TrelloDotNet.Model
         /// </summary>
         /// <param name="message">Message to display to user</param>
         /// <param name="dataSentToTrello">The URI Payload used to call Trello</param>
-        public TrelloApiException(string message, string dataSentToTrello) : base(message)
+        /// <param name="errorCode">Underlying HTTP Status code</param>
+        public TrelloApiException(string message, string dataSentToTrello, HttpStatusCode? errorCode = null) : base(message)
         {
             DataSentToTrello = dataSentToTrello;
+            ErrorCode = errorCode;
         }
     }
 }